requirements.txt
Flask==2.3.3
azure-storage-blob==12.27.0
python-dotenv==1.0.0